% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editset.R
\name{adddummies}
\alias{adddummies}
\title{Add dummy variable to the data.frames, these are needed for errorlocations etc.}
\usage{
adddummies(E, dat)
}
\arguments{
\item{E}{editset}

\item{dat}{data.frame}
}
\value{
data.frame with dummy variables added
}
\description{
Add dummy variable to the data.frames, these are needed for errorlocations etc.
}
\keyword{internal}

