\name{as.character.editmatrix}
\alias{as.character.editmatrix}
\title{Coerce an editmatrix to a \code{character} vector}
\usage{
  \method{as.character}{editmatrix} (x, ...)
}
\arguments{
  \item{x}{editmatrix object to be printed}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
  Derives readable editrules from an editmatrix.
}

