\name{as.character.editset}
\alias{as.character.editset}
\title{Convert an editset to character}
\usage{
  \method{as.character}{editset} (x, datamodel = TRUE,
    useIf = TRUE, dummies = FALSE, ...)
}
\arguments{
  \item{x}{an \code{\link{editset}}}

  \item{datamodel}{include datamodel?}

  \item{useIf}{return vectorized version?}

  \item{dummies}{return datamodel for dummy variables?}

  \item{...}{arguments to be passed to or from other
  methods}
}
\description{
  Convert an editset to character
}

