\name{as.data.frame.editmatrix}
\alias{as.data.frame.editmatrix}
\title{Coerce an editmatrix to a data...}
\usage{\method{as.data.frame}{editmatrix}(x, ...)
}
\description{Coerce an editmatrix to a \code{data.frame}}
\details{Coerces an editmatrix to a \code{data.frame}. Useful for storing/viewing the matrix representation of editrules.}
\value{data.frame with the coefficient matrix representation of \code{x}, a C column and a operator column.}
\arguments{\item{x}{editmatrix object}
\item{...}{further arguments passed to or from other methods.}
}

