% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editset.R
\name{as.editset}
\alias{as.editset}
\title{Coerce x to an editset}
\usage{
as.editset(x, ...)
}
\arguments{
\item{x}{object or vector to be coerced to an editset}

\item{...}{extra parameters that will be passed to \code{as.character}, if necessary}
}
\description{
\code{x} may be an editset, editmatrix, editarray or character vector
}

