\name{as.expression.editmatrix}
\alias{as.expression.editmatrix}
\title{Get expression representation of editmatrix...}
\usage{\method{as.expression}{editmatrix}(x, ...)
}
\description{Get expression representation of editmatrix}
\arguments{\item{x}{editmatrix object to be parsed}
\item{...}{further arguments passed to or from other methods.}
}

