\name{as.lp.mip}
\alias{as.lp.mip}
\title{Coerces a \code{mip} object into an lpsolve object}
\usage{
as.lp.mip(mip)
}
\arguments{
  \item{mip}{object of type \code{mip}.}
}
\description{
\code{as.lp.mip} transforms a mip object into a lpSolveApi
object.
}
\seealso{
\code{\link{as.mip}}, \code{\link{make.lp}}
}

