\name{as.matrix.editarray}
\alias{as.matrix.editarray}
\title{Coerce an editarray to a boolean matrix}
\usage{
  \method{as.matrix}{editarray} (x, ...)
}
\arguments{
  \item{x}{editarray object}

  \item{...}{further arguments to be passed to or from
  other methods}
}
\value{
  boolean matrix of editarray.
}
\description{
  An \code{\link{editarray}} is a boolean array with extra
  attributes. The boolean array may be extracted with this
  method.
}

