% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errorLocalizer_mip.R
\name{asLevels}
\alias{asLevels}
\title{Transform a found solution into a categorical record}
\usage{
asLevels(x)
}
\description{
Transform a found solution into a categorical record
}
\keyword{internal}

