\name{condition}
\alias{condition}
\title{Get condition matrix from an editset.}
\usage{
  condition(E)
}
\arguments{
  \item{E}{an \code{\link{editset}}}
}
\value{
  an \code{\link{editmatrix}}, holding conditions under
  which the editset is relevant.
}
\description{
  Get condition matrix from an editset.
}
\seealso{
  \code{\link{disjunct}}, \code{\link{separate}},
  \code{\link{editset}}
}

