% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contains.R
\name{contains.boolmat}
\alias{contains.boolmat}
\title{Determine if a boolean matrix contains \code{var}}
\usage{
\method{contains}{boolmat}(A, ind, var)
}
\arguments{
\item{A}{array}

\item{ind}{index}

\item{var}{variable name}
}
\description{
Determine if a boolean matrix contains \code{var}
}
\keyword{internal}

