% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicated.R
\name{duplicated.editarray}
\alias{duplicated.editarray}
\title{Check for duplicate edit rules}
\usage{
\method{duplicated}{editarray}(x, ...)
}
\arguments{
\item{x}{a \code{\link{editarray}}}

\item{...}{other parameters to be passed to or from other methods.}
}
\description{
Check for duplicate edit rules
}
\keyword{internal}

