\name{duplicated.editmatrix}
\alias{duplicated.editmatrix}
\title{Check for duplicate edit rules...}
\usage{duplicated.editmatrix(x, ...)
}
\description{Check for duplicate edit rules}
\value{logical vector}
\arguments{\item{x}{an \code{\link{editmatrix}}}
\item{...}{options to be passed to other methods}
}

