\name{editmatrix-subscript}
\alias{[.editmatrix}
\title{Row index operator for editmatrix.}
\usage{`[.editmatrix`(x,i,j,...)}
\description{Row index operator for \code{editmatrix}.}
\details{Use this operator to select edits from an editmatrix object.}
\arguments{\item{x}{an object of class \code{\link{editmatrix}}}
\item{i}{the row index in the edit matrix}
\item{j}{the column index in the edit matrix}
\item{...}{arguments to be passed to other methods. Currently ignored.}
}

