% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editAttr.R
\name{editnames}
\alias{editnames}
\title{Names of edits}
\usage{
editnames(E)
}
\arguments{
\item{E}{\code{\link{editset}}, \code{\link{editarray}} or \code{\link{editmatrix}}}
}
\description{
Retrieve edit names from editset, -array or -matrix
}

