\name{errorLocalizer.editmatrix}
\alias{errorLocalizer.editmatrix}
\title{Localize errors in numerical data...}
\usage{\method{errorLocalizer}{editmatrix}(E, x, weight=rep(1, length(x)), ...)
}
\description{Localize errors in numerical data}
\arguments{\item{E}{Object of class \code{\link{editmatrix}}}
\item{x}{Data record, in the form of a named numeric vector.}
\item{weight}{Weight vector, of the same length of \code{x}}
\item{...}{arguments to be passed to other methods.}
}

