% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editarrayAttr.R
\name{getArr}
\alias{getArr}
\title{Get named logical array from editarray}
\usage{
getArr(E)
}
\arguments{
\item{E}{\code{\link{editarray}}}
}
\value{
logical array
}
\description{
Get named logical array from editarray
}
\keyword{internal}

