% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getlevels}
\alias{getlevels}
\title{retrieve level names from editarray}
\usage{
getlevels(E)
}
\arguments{
\item{editarray}{\code{\link{editarray}}}
}
\value{
character vector
}
\description{
retrieve level names from editarray
}
\keyword{internal}

