% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editarrayAttr.R
\name{getnames}
\alias{getnames}
\title{retrieve edit names from editarray}
\usage{
getnames(E)
}
\arguments{
\item{E}{\code{\link{editarray}}}
}
\value{
character vector
}
\description{
retrieve edit names from editarray
}
\keyword{internal}

