% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editarrayAttr.R
\name{getSep}
\alias{getSep}
\title{get seprator used to seperate variables from levels in editarray}
\usage{
getSep(E)
}
\arguments{
\item{E}{\code{\link{editarray}}}
}
\value{
character
}
\description{
get seprator used to seperate variables from levels in editarray
}
\keyword{internal}

