% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getVars.R
\name{getVars.cateditmatrix}
\alias{getVars.cateditmatrix}
\title{Returns the variable names of an (in)equality \code{editmatrix} E}
\usage{
\method{getVars}{cateditmatrix}(E, type = c("uniquevar", "colnames", "var",
  "cat"), ...)
}
\arguments{
\item{type}{should unique variable names, colnames, all variable names or category names be extracted?}
}
\description{
Returns the variable names of an (in)equality \code{editmatrix} E
}
\keyword{internal}

