% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getVars.R
\name{getVars.editarray}
\alias{getVars.editarray}
\title{get variable names in editarray}
\usage{
\method{getVars}{editarray}(E, type = "cat", ...)
}
\description{
get variable names in editarray
}
\keyword{internal}

