% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getVars.editmatrix}
\alias{getVars.editmatrix}
\title{Returns the variable names of an (in)equality \code{editmatrix} E}
\usage{
\method{getVars}{editmatrix}(E, ...)
}
\description{
Returns the variable names of an (in)equality \code{editmatrix} E
}
\keyword{internal}

