% Generated by roxygen2 (4.0.1): do not edit by hand
\name{indFromArray}
\alias{indFromArray}
\title{Compute index from array part of editarray}
\usage{
indFromArray(A, sep)
}
\arguments{
\item{A}{boolean array}

\item{sep}{separator}
}
\description{
Compute index from array part of editarray
}
\keyword{internal}

