\name{is.cateditmatrix}
\alias{is.cateditmatrix}
\title{Check if object is a cateditmatrix}
\usage{
  is.cateditmatrix(x)
}
\arguments{
  \item{x}{object to be checked}
}
\value{
  TRUE if \code{x} is an \code{cateditmatrix}
}
\description{
  Please note that a cateditmatrix is also an editmatrix
}
\seealso{
  \code{\link{cateditmatrix}}
}

