\name{is.editarray}
\alias{is.editarray}
\title{check if an object is an editarray}
\usage{
  is.editarray(x)
}
\arguments{
  \item{x}{object to be checked}
}
\value{
  \code{TRUE} if \code{x} is an \code{\link{editarray}}
}
\description{
  check if an object is an editarray
}
\seealso{
  \code{\link{editarray}}, \code{\link{is.editmatrix}}
}

