\name{is.editmatrix}
\alias{is.editmatrix}
\title{Check if object is an editmatrix}
\usage{
  is.editmatrix(x)
}
\arguments{
  \item{x}{object to be checked}
}
\value{
  TRUE if \code{x} is an \code{editmatrix}
}
\description{
  Check if object is an editmatrix
}
\seealso{
  \code{\link{editmatrix}}
}

