\name{is.editset}
\alias{is.editset}
\title{Check if an object is an \code{\link{editset}}}
\usage{
  is.editset(x)
}
\arguments{
  \item{x}{object to be checked}
}
\value{
  \code{TRUE} if \code{x} is an \code{\link{editset}}
}
\description{
  Check if an object is an \code{\link{editset}}
}
\seealso{
  \code{\link{editarray}}, \code{\link{is.editmatrix}}
}

