\name{isObviouslyRedundant.editmatrix}
\alias{isObviouslyRedundant.editmatrix}
\title{Redundancy check, editmatrix method...}
\usage{\method{isObviouslyRedundant}{editmatrix}(E, ...)
}
\description{Redundancy check, \code{editmatrix} method}
\seealso{\code{\link{isObviouslyRedundant}}, \code{\link{isObviouslyRedundant.matrix}}}
\arguments{\item{E}{Augmented matrix [A|b] or \code{\link{editmatrix}}}
\item{...}{parameters to be passed to other methods. Currently, only \code{tol} is implemented (see \code{\link{isObviouslyRedundant.matrix}}).}
}

