\name{neweditarray}
\alias{neweditarray}
\title{editarray: logical array where every column corresponds to one
level of one variable. Every row is an edit. Every edit denotes
a *forbidden* combination.}
\usage{
  neweditarray(E, ind, sep, names = NULL,
    levels = colnames(E), ...)
}
\description{
  editarray: logical array where every column corresponds
  to one level of one variable. Every row is an edit. Every
  edit denotes a *forbidden* combination.
}
\keyword{internal}

