\name{numcat}
\alias{numcat}
\title{return vector with number of categories for each variable in editarray}
\usage{
  numcat(E)
}
\arguments{
  \item{E}{object of class \code{\link{editarray}}}
}
\value{
  a vector of \code{length(getVar(E))} with the number of
  categories for every variable in E.
}
\description{
  return vector with number of categories for each variable
  in editarray
}

