% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parseNum.R
\name{parseNum}
\alias{parseNum}
\title{Parse a numerical edit expression}
\usage{
parseNum(e)
}
\arguments{
\item{e}{a valid R expression}
}
\description{
Parse a numerical edit expression into a named \code{numeric}.
The \code{names} are the names of the variables
}
\keyword{internal}

