\name{plot.errorLocation}
\alias{plot.errorLocation}
\title{plot errorLocation object}
\usage{
  \method{plot}{errorLocation} (x, topn = min(10,
  ncol(x$adapt)), ...)
}
\arguments{
  \item{x}{errorLocation object}

  \item{topn}{Number of variables to show in 'errors per
  variable plot'. Only the top-n are are shown. By default
  the top-20 variables with the most errors are shown.}

  \item{...}{other arguments that will be transferred to
  \code{barplot}}
}
\description{
  Gives an overview of results generated by
  \code{link{localizeErrors}}. Axes in scatterplots are set
  to logarithmic if their scale maxima exceed 50.
}

