\name{plot.violatedEdits}
\alias{plot.violatedEdits}
\title{Plot summary statistics on violatedEdits}
\usage{
  \method{plot}{violatedEdits} (x, topn = min(10, ncol(x)),
  ...)
}
\arguments{
  \item{x}{\code{violatedEdits} object.}

  \item{topn}{Top \code{n} edits to be plotted.}

  \item{...}{parameters passed to \code{barplot} method.}
}
\description{
  Plot summary statistics on violatedEdits
}

