% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.cateditmatrix}
\alias{print.cateditmatrix}
\title{print cateditmatrix}
\usage{
\method{print}{cateditmatrix}(x, textOnly = TRUE, ...)
}
\arguments{
\item{x}{cateditmatrix object to be printed}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print cateditmatrix
}
\keyword{internal}

