\name{print.choicepoint}
\alias{print.choicepoint}
\title{print a choicepoint...}
\usage{\method{print}{choicepoint}(x, ..., VERBOSE=FALSE)
}
\description{print a choicepoint}
\arguments{\item{x}{choicepoint object to be printed}
\item{...}{other parameters passed to print method}
\item{VERBOSE}{should all variables be printed?}
}

