\name{print.editarray}
\alias{print.editarray}
\title{print  editarray}
\usage{
\method{print}{editarray}(x, textOnly = FALSE, ...)
}
\arguments{
  \item{x}{an \code{\link{editarray}}}

  \item{textOnly}{If \code{FALSE}, also print internal
  structure}

  \item{...}{arguments to be passed to or from other
  methods.}
}
\description{
print editarray
}
\keyword{internal}

