% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.editlist}
\alias{print.editlist}
\title{print editset}
\usage{
\method{print}{editlist}(x, ...)
}
\arguments{
\item{x}{editset object to be printed}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print editset
}
\keyword{internal}

