% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.editmatrix}
\alias{print.editmatrix}
\title{print editmatrix}
\usage{
\method{print}{editmatrix}(x, textOnly = FALSE, ...)
}
\arguments{
\item{x}{editmatrix object to be printed}

\item{textOnly}{If \code{FALSE}, also print internal structure}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print editmatrix
}
\keyword{internal}

