\name{removeEmpty}
\alias{removeEmpty}
\title{Reduce an (edit)matrix by removing empty rows and columns}
\usage{
  removeEmpty(E)
}
\arguments{
  \item{E}{an object that is coerceable to a \code{matrix}}
}
\description{
  Reduce an (edit)matrix by removing empty rows and columns
}
\keyword{internal}

