% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editset.R
\name{removeRedundantDummies}
\alias{removeRedundantDummies}
\title{Remove redundant dummy variables}
\usage{
removeRedundantDummies(E, tol = 1e-08)
}
\arguments{
\item{E}{\code{\link{editset}}}

\item{tol}{positive number}
}
\description{
Remove duplicated dummy variables from an editset.
}
\keyword{internal}

