\name{softEdits.cateditmatrix}
\alias{softEdits.cateditmatrix}
\title{Derive editmatrix with soft constraints. This is a utility function that is used for
constructing a mip/lp problem.}
\usage{
  softEdits.cateditmatrix(E, prefix = "delta.", ...)
}
\arguments{
  \item{E}{normalized \code{editmatrix}}

  \item{prefix}{\code{character} used for naming dummy
  variables in matrix.}
}
\description{
  Derive editmatrix with soft constraints. This is a
  utility function that is used for constructing a mip/lp
  problem.
}
\keyword{internal}

