\name{str.editarray}
\alias{str.editarray}
\title{\code{\link{str}} method for editarray object}
\usage{
  \method{str}{editarray} (object, ...)
}
\arguments{
  \item{object}{\code{\link{editarray}} object}

  \item{...}{arguments to pass to other methods}
}
\description{
  \code{\link{str}} method for editarray object
}

