\name{summary.violatedEdits}
\alias{summary.violatedEdits}
\title{Summary statistics on violatedEdits}
\usage{
  \method{summary}{violatedEdits} (object, E = NULL,
    minfreq = 1, ...)
}
\arguments{
  \item{object}{\code{violatedEdits} object}

  \item{E}{optional \code{editmatrix} or \code{editarray}}

  \item{minfreq}{minimum freq for edit to be printed}

  \item{...}{(not used)}
}
\description{
  Summary statistics on violatedEdits
}

