\name{as.character.cateditmatrix}
\alias{as.character.cateditmatrix}
\title{Coerce an cateditmatrix to a \code{character} vector}
\usage{
  \method{as.character}{cateditmatrix} (x, ...)
}
\arguments{
  \item{x}{cateditmatrix object to be printed}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
  Derives readable editrules from an editmatrix.
}
\keyword{internal}

