\name{as.character.editmatrix}
\alias{as.character.editmatrix}
\title{Get character representation of editmatrix...}
\usage{\method{as.character}{editmatrix}(x, ...)
}
\description{Get character representation of editmatrix}
\arguments{\item{x}{editmatrix object to be printed}
\item{...}{further arguments passed to or from other methods.}
}

