\name{as.editmatrix}
\alias{as.editmatrix}
\title{Transform a matrix into an edit matrix...}
\usage{as.editmatrix(x)
}
\description{Transform a matrix into an edit matrix}
\details{The columns of the matrix
are the variables and the rows are the edit rules (contraints).}
\seealso{\code{\link{editmatrix}}}
\value{an object of class \code{editmatrix}.}
\arguments{\item{x}{object to be transformed into an \code{\link{editmatrix}}}
}

