\name{as.expression.editmatrix}
\alias{as.expression.editmatrix}
\title{Coerce an editmatrix to R expressions...}
\usage{\method{as.expression}{editmatrix}(x, ...)
}
\description{Coerce an editmatrix to R expressions}
\details{Generates an R \code{expression} vector that can be used to check data using \code{\link{eval}}.}
\arguments{\item{x}{editmatrix object to be parsed}
\item{...}{further arguments passed to or from other methods.}
}

