\name{as.matrix.editmatrix}
\alias{as.matrix.editmatrix}
\title{Convert an editmatrix to a normal matrix...}
\usage{\method{as.matrix}{editmatrix}(x, ...)
}
\description{Convert an editmatrix to a normal matrix}
\details{An \code{editmatrix} is a matrix and can be used as such, but it has extra attributes.
In some case it is preferable to convert the editmatrix to a normal matrix.}
\value{matrix equal to editmatrix}
\arguments{\item{x}{editmatrix object}
\item{...}{further arguments passed to or from other methods.}
}

