\name{contains.boolmat}
\alias{contains.boolmat}
\title{determine if a boolean matrix contains var}
\usage{
  contains.boolmat(A, ind, var)
}
\arguments{
  \item{A}{array}

  \item{ind}{index}

  \item{var}{variable name}
}
\description{
  determine if a boolean matrix contains var
}
\keyword{internal}

