% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{contains.boolmat}
\alias{contains.boolmat}
\title{Determine if a boolean matrix contains \code{var}}
\usage{
\method{contains}{boolmat}(E, var = NULL, ind, ...)
}
\arguments{
\item{E}{array}

\item{var}{variable name}

\item{ind}{index}
}
\description{
Determine if a boolean matrix contains \code{var}
}
\keyword{internal}
