\name{contains.boolmat}
\alias{contains.boolmat}
\title{Determine if a boolean matrix contains \code{var}}
\usage{
  contains.boolmat(A, ind, var)
}
\arguments{
  \item{A}{array}

  \item{ind}{index}

  \item{var}{variable name}
}
\description{
  Determine if a boolean matrix contains \code{var}
}
\keyword{internal}

