\name{cp.editmatrix}
\alias{cp.editmatrix}
\title{Deprecated error localization function.}
\usage{cp.editmatrix(E, x, ...)
}
\description{Deprecated error localization function.}
\details{This function is replaced by S3 generic \code{\link{errorLocalizer}}.}
\arguments{\item{E}{editmatrix}
\item{x}{record}
\item{...}{Arguments to be passed to \code{\link{errorLocalizer}}}
}

