\name{duplicated.editmatrix}
\alias{duplicated.editmatrix}
\title{Check for duplicate edit rules}
\usage{
  \method{duplicated}{editmatrix} (x, ...)
}
\arguments{
  \item{x}{an \code{\link{editmatrix}}}

  \item{...}{options to be passed to other methods}
}
\value{
  logical vector
}
\description{
  Check for duplicate edit rules
}
\keyword{internal}

