\name{editnames}
\alias{editnames}
\title{Names of edits
Retrieve edit names from editset, -array or -matrix}
\usage{
  editnames(E)
}
\arguments{
  \item{E}{\code{\link{editset}}, \code{\link{editarray}}
  or \code{\link{editmatrix}}}
}
\description{
  Names of edits Retrieve edit names from editset, -array
  or -matrix
}

