% Generated by roxygen2 (4.0.1): do not edit by hand
\name{editnames}
\alias{editnames}
\title{Names of edits}
\usage{
editnames(E)
}
\arguments{
\item{E}{\code{\link{editset}}, \code{\link{editarray}} or \code{\link{editmatrix}}}
}
\description{
Retrieve edit names from editset, -array or -matrix
}

