\name{editset}
\alias{editset}
\title{Create an editset which can contain a mix of categorical, numerical and mixededits}
\usage{
  editset(editrules)
}
\arguments{
  \item{editrules}{\code{data.frame} with (in)equalities
  written in R syntax, see details for description or
  alternatively a \code{character} or \code{expression}
  with (in)equalities written in R syntax}
}
\description{
  Create an editset which can contain a mix of categorical,
  numerical and mixededits
}

