\name{eliminate}
\alias{eliminate}
\title{eliminate a variable from a set of edit rules...}
\usage{eliminate(E, var, ...)
}
\description{eliminate a variable from a set of edit rules}
\arguments{\item{E}{editmatrix}
\item{var}{name of variable to be eliminated}
\item{...}{argumemts to be passed to or from other methods}
}

