% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/editarrayAttr.R
\name{getlevels}
\alias{getlevels}
\title{retrieve level names from editarray}
\usage{
getlevels(E)
}
\arguments{
\item{editarray}{\code{\link{editarray}}}
}
\value{
character vector
}
\description{
retrieve level names from editarray
}
\keyword{internal}

