\name{getSep}
\alias{getSep}
\title{get seprator used to seperate variables from levels in editarray}
\usage{
  getSep(E)
}
\arguments{
  \item{E}{\code{\link{editarray}}}
}
\value{
  character
}
\description{
  get seprator used to seperate variables from levels in
  editarray
}
\keyword{internal}

