\name{is.editarray}
\alias{is.editarray}
\title{Check if an object is an \code{\link{editarray}}}
\usage{
  is.editarray(x)
}
\arguments{
  \item{x}{object to be checked}
}
\value{
  \code{TRUE} if \code{x} is an \code{\link{editarray}}
}
\description{
  Check if an object is an \code{\link{editarray}}
}
\seealso{
  \code{\link{is.editset}}, \code{\link{is.editmatrix}}
}

