\name{is.editrules}
\alias{is.editarray}
\alias{is.editmatrix}
\alias{is.editrules}
\alias{is.editset}
\title{Check object class}
\usage{
  is.editset(x)

  is.editmatrix(x)

  is.editarray(x)
}
\arguments{
  \item{x}{object to be checked}
}
\value{
  \code{logical}
}
\description{
  Check object class
}

