\name{isNormalized}
\alias{isNormalized}
\title{Check if an editmatrix is normalized}
\usage{
  isNormalized(E)
}
\arguments{
  \item{E}{\code{\link{editmatrix}}}
}
\value{
  TRUE when all comparison operators of \code{E} are in
  \{\code{<,<=,==}\}
}
\description{
  Check if an editmatrix is normalized
}
\seealso{
  \code{\link{editmatrix}}
}

