\name{nedits}
\alias{nedits}
\title{Number of edits
Count the number of edits in a collection of edits.}
\usage{
  nedits(E)
}
\arguments{
  \item{E}{\code{\link{editset}}, \code{\link{editarray}}
  or \code{\link{editmatrix}}}
}
\description{
  Number of edits Count the number of edits in a collection
  of edits.
}

