% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parseCatEdit}
\alias{parseCatEdit}
\title{parse categorial edit}
\usage{
parseCatEdit(el)
}
\arguments{
\item{e}{\code{expression} with a single edit}
}
\value{
named \code{numeric} with coefficients
}
\description{
parse categorial edit
}
\keyword{internal}

