\name{print.backtracker}
\alias{print.backtracker}
\title{print a backtracker...}
\usage{\method{print}{backtracker}(x, ..., VERBOSE=FALSE)
}
\description{print a backtracker}
\arguments{\item{x}{backtracker object to be printed}
\item{...}{other parameters passed to print method}
\item{VERBOSE}{should all variables be printed?}
}

