% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backtracker.R
\name{print.backtracker}
\alias{print.backtracker}
\title{print a backtracker}
\usage{
\method{print}{backtracker}(x, ..., VERBOSE = FALSE)
}
\arguments{
\item{x}{backtracker object to be printed}

\item{...}{other parameters passed to print method}

\item{VERBOSE}{should all variables be printed?}
}
\description{
print a backtracker
}
\keyword{internal}

