% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.editarray}
\alias{print.editarray}
\title{print  editarray}
\usage{
\method{print}{editarray}(x, textOnly = FALSE, ...)
}
\arguments{
\item{x}{an \code{\link{editarray}}}

\item{textOnly}{If \code{FALSE}, also print internal structure}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
print  editarray
}
\keyword{internal}

