\name{print.editmatrix}
\alias{print.editmatrix}
\title{print editmatrix}
\usage{
  \method{print}{editmatrix} (x, ...)
}
\arguments{
  \item{x}{editmatrix object to be printed}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
  print editmatrix
}
\keyword{internal}

