\name{print.editmatrix}
\alias{print.editmatrix}
\title{print editmatrix}
\usage{
\method{print}{editmatrix}(x, textOnly = FALSE, ...)
}
\arguments{
  \item{x}{editmatrix object to be printed}

  \item{textOnly}{If \code{FALSE}, also print internal
  structure}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
print editmatrix
}
\keyword{internal}

