% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.editset}
\alias{print.editset}
\title{print editset}
\usage{
\method{print}{editset}(x, ...)
}
\arguments{
\item{x}{editset object to be printed}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print editset
}
\keyword{internal}

