% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.errorLocation}
\alias{print.errorLocation}
\title{Print object of class errorLocation}
\usage{
\method{print}{errorLocation}(x, ...)
}
\arguments{
\item{x}{object of class errorLocation}

\item{...}{arguments to be passed to other methods}
}
\description{
Print object of class errorLocation
}
\keyword{internal}

