\name{print.errorLocation}
\alias{print.errorLocation}
\title{Print object of class errorLocation}
\usage{
  \method{print}{errorLocation} (x, ...)
}
\arguments{
  \item{x}{object of class errorLocation}

  \item{...}{arguments to be passed to other methods}
}
\description{
  Print object of class errorLocation
}
\keyword{internal}

