\name{str.editmatrix}
\alias{str.editmatrix}
\title{\link{str} method for editmatrix object...}
\usage{str.editmatrix(object, ...)
}
\description{\code{\link{str}} method for editmatrix object}
\arguments{\item{object}{\code{\link{editmatrix}} object}
\item{...}{methods to pass to other methods}
}

