\name{str.editmatrix}
\alias{str.editmatrix}
\title{\code{\link{str}} method for editmatrix object}
\usage{
  \method{str}{editmatrix} (object, ...)
}
\arguments{
  \item{object}{\code{\link{editmatrix}} object}

  \item{...}{arguments to pass to other methods}
}
\description{
  \code{\link{str}} method for editmatrix object
}

