\name{[.editmatrix}
\alias{[.editarray}
\alias{[.editmatrix}
\title{Row index operator for \code{editmatrix}}
\usage{
  \method{[}{editmatrix} (x, i, j, ...)

  \method{[}{editarray} (x, i, j, ...)
}
\arguments{
  \item{x}{an object of class \code{\link{editmatrix}} or
  \code{\link{editarray}}}

  \item{i}{the row index in the edit matrix}

  \item{j}{the column index in the edit matrix}

  \item{...}{arguments to be passed to other methods.
  Currently ignored.}
}
\description{
  Use this operator to select edits from an editmatrix or
  editarray object.

  Row index operator for \code{editarray}
}

